In atm merge, make sure arrays for other components are initialized
xao_am, l2x_am, i2x_am, and o2x_am are only set using GetDouble if
the corresponding model is present.  For at least o2x_am, it was
still used in a merge even if not present.  So 0 those arrays
if the corresponding moab app id is not set.

This zero's the Faoo_h2otemp field for the moab coupler.