Add newer coupling capability namelist and flags
Add newer coupling capability namelist and flags to driver-moab.
This includes changes for one-way ocean-to-land, iac and glacier-wave coupling
Also adds abilit to add more coupling fields for these configs.

This adds some logical flags that are now expected in other components
and in the namelist.  Does not yet implement the new couplings.

Need to remove build errors on master associated with ocean-to-land additions.

[BFB]