Fix undefined behavior in MPAS registry parser
With `gcc` v13.3.0, I'm seeing unparsable `if (Active .or. analysisModeActive)` clauses in the file `components/mpas-ocean/src/inc/structs_and_variables.inc` generated by the registry parser.

The issue appears to be that the registry parser is reading from and writing to the same string in statements like:
```
sprintf(out_packages, "%s;%s", out_packages, token);
```
This is not allowed, as it can lead to race conditions and bad output like I am seeing. This fix avoids this undefined behavior when appending to `out_packages` in the registry parser.